+++ server/main/passport/index.js
@@ -1,4 +1,3 @@
-const passport = require('passport');
 const customGithub = require('./customStrategy');
 
 module.exports = passport => {
